feat: Added Docker Compose configuration for PostgreSQL, PgAdmin, and Redis Cache
Created the docker-compose.yml file to set up services for PostgreSQL Database, PgAdmin, and Redis Cache. The services use the latest versions of the respective images and define environment variables to configure them.

Added a .env file to set environment variables for the services, such as PostgreSQL host, port, database name, user, and password, as well as Redis host and port, and PgAdmin default email, password, and port.

This commit enables easier setup and deployment of the required services using Docker Compose.